Updated Entity.h

+++ Entity.h
@@ -8,13 +8,13 @@
 class Entity {
     protected:
         Vector2 position;
-        size_t hp;
-        const size_t maxHP;
+        int hp;
+        const int maxHP;
         std::string name;
         std::unordered_map<EffectType, Effect> effects;
         int def;
 
-        Entity(std::string Name, int startX, int startY, size_t health, size_t MaxHP, int def);
+        Entity(std::string Name, int startX, int startY, int health, int MaxHP, int def);
         Entity();
     public:
         virtual ~Entity() = default;
@@ -22,7 +22,7 @@ class Entity {
         virtual void update(int pX,int pY) = 0;
         void move(int dx, int dy);
         void takeDamage(int dmg);
-        void Healing(size_t heal);
+        void Healing(int heal);
         virtual int Damage() const = 0;
         virtual int Defense() const = 0;
         virtual void Effect_Action() = 0;
@@ -30,8 +30,8 @@ class Entity {
 
         int PosX() const;
         int posY() const;
-        size_t HP() const;
-        size_t MaxHP() const;
+        int HP() const;
+        int MaxHP() const;
         bool isAlive() const;
         std::string Name() const;
         virtual char getSymbol() const = 0;
